Use the Spring Initializr
I rewrote the document and the code (mostly changing package names) to use the Spring Initializr. This particular example involved modifying both the Maven and the Gradle build files, which added quite a bit to the document.
