+++ complete/application/src/main/java/com/example/multimodule/application/DemoApplication.java
@@ -1,12 +1,12 @@
-package hello.app;
+package com.example.multimodule.application;
 
-import hello.service.MyService;
+import com.example.multimodule.service.MyService;
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RestController;
 
-@SpringBootApplication(scanBasePackages = "hello")
+@SpringBootApplication(scanBasePackages = "com.example.multimodule")
 @RestController
 public class DemoApplication {
