+++ complete/application/src/main/resources/application.properties
@@ -1,1 +1,1 @@
-service.message=Hello World
+service.message=Hello, World
