+++ complete/application/src/test/java/com/example/multimodule/application/DemoApplicationTest.java
@@ -1,28 +1,25 @@
-package hello.service;
+package com.example.multimodule.application;
 
-import static org.assertj.core.api.Assertions.*;
+import static org.assertj.core.api.Assertions.assertThat;
 
 import org.junit.Test;
 import org.junit.runner.RunWith;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
 import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.test.context.junit4.SpringRunner;
 
+import com.example.multimodule.service.MyService;
+
 @RunWith(SpringRunner.class)
-@SpringBootTest("service.message=Hello")
-public class MyServiceTest {
+@SpringBootTest
+public class DemoApplicationTest {
 
     @Autowired
     private MyService myService;
 
     @Test
     public void contextLoads() {
-        assertThat(myService.message()).isNotNull();
-    }
-
-    @SpringBootApplication
-    static class TestConfiguration {
+        assertThat(myService).isNotNull();
     }
 
 }
