+++ complete/library/build.gradle
@@ -1,27 +1,26 @@
-buildscript {
-    repositories { mavenCentral() }
+plugins {
+	id 'org.springframework.boot' version '2.1.8.RELEASE'
+	id 'io.spring.dependency-management' version '1.0.8.RELEASE'
+	id 'java'
 }
 
-plugins { id "io.spring.dependency-management" version "1.0.5.RELEASE" }
+group = 'com.example'
+version = '0.0.1-SNAPSHOT'
+sourceCompatibility = '1.8'
 
-ext { springBootVersion = '2.1.6.RELEASE' }
+repositories {
+	mavenCentral()
+}
 
-apply plugin: 'java'
-apply plugin: 'eclipse'
+bootJar {
+	enabled = false
+}
 
 jar {
-    baseName = 'gs-multi-module-library'
-    version = '0.0.1-SNAPSHOT'
+	enabled = true
 }
-sourceCompatibility = 1.8
-
-repositories { mavenCentral() }
 
 dependencies {
-    compile('org.springframework.boot:spring-boot-starter')
-    testCompile('org.springframework.boot:spring-boot-starter-test')
-}
-
-dependencyManagement {
-    imports { mavenBom("org.springframework.boot:spring-boot-dependencies:${springBootVersion}") }
+	implementation 'org.springframework.boot:spring-boot-starter'
+	testImplementation 'org.springframework.boot:spring-boot-starter-test'
 }
