+++ complete/library/src/main/java/com/example/multimodule/service/MyService.java
@@ -1,4 +1,4 @@
-package hello.service;
+package com.example.multimodule.service;
 
 import org.springframework.boot.context.properties.EnableConfigurationProperties;
 import org.springframework.stereotype.Service;
