+++ complete/library/src/main/java/com/example/multimodule/service/ServiceProperties.java
@@ -1,4 +1,4 @@
-package hello.service;
+package com.example.multimodule.service;
 
 import org.springframework.boot.context.properties.ConfigurationProperties;
