+++ complete/library/src/test/java/com/example/multimodule/service/MyServiceTest.java
@@ -1,25 +1,28 @@
-package hello.app;
+package com.example.multimodule.service;
 
-import static org.assertj.core.api.Assertions.assertThat;
+import static org.assertj.core.api.Assertions.*;
 
 import org.junit.Test;
 import org.junit.runner.RunWith;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
 import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.test.context.junit4.SpringRunner;
 
-import hello.service.MyService;
-
 @RunWith(SpringRunner.class)
-@SpringBootTest
-public class DemoApplicationTest {
+@SpringBootTest("service.message=Hello")
+public class MyServiceTest {
 
     @Autowired
     private MyService myService;
 
     @Test
     public void contextLoads() {
-        assertThat(myService).isNotNull();
+        assertThat(myService.message()).isNotNull();
+    }
+
+    @SpringBootApplication
+    static class TestConfiguration {
     }
 
 }
